feat: add address mode calculation & complete pass one code; start pass two code

+++ chasm/opcodes.swift
@@ -16,8 +16,8 @@ enum AddressingMode: Decodable {
     case immediate		// immediate operand (e.g. LDA #$FF)
     case zeroPage		// one byte address operand (e.g. LDA $42)
     case zeroPageX		// adds X to zero-page address (e.g. LDA $42,X)
-    case zeroPageY		// adds Y to zero-page address (e.g. LDA $42,Y)
-    case indirect		// absolute indirect, JMP only (e.g. JMP $fffe)
+    case zeroPageY		// adds Y to zero-page address (e.g. LDX $42,Y)
+    case indirect		// absolute indirect, JMP only (e.g. JMP ($fffe))
     case indexedIndirect // index ZP addres with X, fetch address from there (e.g. LDA ($20,X))
     case indirectIndexed // fetch 16-bit address from ZP, then add & to it (e.g. LDA($20), Y)
     case relative		// branch opcodes only (e.g. BEQ .label, where label is -128 to +127 one byte signed offset)
@@ -26,7 +26,7 @@ enum AddressingMode: Decodable {
     case absoluteX		// adds X to 16-bit address op (e.g. LDA $4200,X)
     case absoluteY		// adds Y to 16-bit address op (e.g. LDA $4200,Y)
     
-    var byteSize: Int {
+    var byteSize: UInt16 {
         switch self {
         case .implied, .accumulator: return 1
         case .immediate, .zeroPage, .zeroPageX, .zeroPageY, .indexedIndirect, .indirectIndexed, .relative: return 2
